all the other path renames post move
