+++ src/vs/workbench/contrib/preferences/browser/settingsEditorSettingIndicators.ts
@@ -21,7 +21,7 @@ import { IUserDataSyncEnablementService } from 'vs/platform/userDataSync/common/
 import { SettingsTreeSettingElement } from 'vs/workbench/contrib/preferences/browser/settingsTreeModels';
 import { POLICY_SETTING_TAG } from 'vs/workbench/contrib/preferences/common/preferences';
 import { IWorkbenchConfigurationService } from 'vs/workbench/services/configuration/common/configuration';
-import { IHoverOptions, IHoverService, IHoverWidget } from 'vs/workbench/services/hover/browser/hover';
+import { IHoverOptions, IHoverService, IHoverWidget } from 'vs/platform/hover/browser/hover';
 
 const $ = DOM.$;
