+++ src/vs/workbench/contrib/terminal/browser/terminalTooltip.ts
@@ -7,7 +7,7 @@ import { localize } from 'vs/nls';
 import { ITerminalInstance } from 'vs/workbench/contrib/terminal/browser/terminal';
 import { TerminalCapability } from 'vs/platform/terminal/common/capabilities/capabilities';
 import { asArray } from 'vs/base/common/arrays';
-import { IHoverAction } from 'vs/workbench/services/hover/browser/hover';
+import { IHoverAction } from 'vs/platform/hover/browser/hover';
 import { MarkdownString } from 'vs/base/common/htmlContent';
 
 export function getInstanceHoverInfo(instance: ITerminalInstance): { content: MarkdownString; actions: IHoverAction[] } {
