+++ src/vs/workbench/contrib/terminalContrib/links/browser/links.ts
@@ -5,7 +5,7 @@
 
 import type { IBufferLine, IBufferRange, Terminal } from '@xterm/xterm';
 import { URI } from 'vs/base/common/uri';
-import { IHoverAction } from 'vs/workbench/services/hover/browser/hover';
+import { IHoverAction } from 'vs/platform/hover/browser/hover';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { ITerminalProcessManager } from 'vs/workbench/contrib/terminal/common/terminal';
 import { IParsedLink } from 'vs/workbench/contrib/terminalContrib/links/browser/terminalLinkParsing';
