+++ src/vs/workbench/contrib/testing/browser/testCoverageBars.ts
@@ -19,7 +19,7 @@ import { TestingConfigKeys, TestingDisplayedCoveragePercent, getTestingConfigura
 import { AbstractFileCoverage } from 'vs/workbench/contrib/testing/common/testCoverage';
 import { ITestCoverageService } from 'vs/workbench/contrib/testing/common/testCoverageService';
 import { ICoveredCount } from 'vs/workbench/contrib/testing/common/testTypes';
-import { IHoverService } from 'vs/workbench/services/hover/browser/hover';
+import { IHoverService } from 'vs/platform/hover/browser/hover';
 
 export interface TestCoverageBarsOptions {
 	/**
